backend/requirements.txt
# annotated-types==0.7.0
# anyio==4.10.0
# click==8.2.1
# fastapi==0.116.1
# greenlet==3.2.4
# h11==0.16.0
# httptools==0.6.4
# idna==3.10
# psycopg2-binary==2.9.10
# pydantic==2.11.7
# pydantic_core==2.33.2
# python-dotenv==1.1.1
# PyYAML==6.0.2
# sniffio==1.3.1
# SQLAlchemy==2.0.43
# starlette==0.47.3
# typing-inspection==0.4.1
# typing_extensions==4.15.0
# uvicorn==0.35.0
# uvloop==0.21.0
# watchfiles==1.1.0
# websockets==15.0.1
# passlib[bcrypt]==1.8.2
# python-jose==3.3.0
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic
